Added UI Manager.
