+++ WakaSkies.Desktop/UI/Myra Projects/NoticeWindow.xmmp
@@ -1,5 +1,5 @@
 <Project>
-  <Window Title="" Left="568" Top="237" Width="300" Height="150">
+  <Window Title="" Left="568" Top="237" Width="400" Height="200">
     <Panel>
       <Label Wrap="True" HorizontalAlignment="Center" Id="contentLabel" />
       <TextButton Text="Close" Margin="5" Padding="5" HorizontalAlignment="Right" VerticalAlignment="Bottom" Id="closeButton" />
